added endpoints

+++ client/src/API.js
@@ -28,11 +28,13 @@ export const dropDatabase = async (data) => {
 
 export const createTable = async (data) => {
 	const name = data?.name;
-	if (name) {
+	const dbName = data?.dbName;
+	if (name && dbName) {
 		return await fetch(`${baseURL}table`, {
 			method: 'POST',
 			body: JSON.stringify({
-				name
+				name,
+				dbName
 			}),
 			headers: {
 				'Content-Type': 'application/json'
@@ -43,20 +45,31 @@ export const createTable = async (data) => {
 
 export const dropTable = async (data) => {
 	const name = data?.name;
-	if (name) {
+	const dbName = data?.dbName;
+	if (name && dbName) {
 		return await fetch(`${baseURL}table/${name}`, {
-			method: 'DELETE'
+			method: 'DELETE',
+			body: JSON.stringify({
+				dbName
+			}),
+			headers: {
+				'Content-Type': 'application/json'
+			}
 		});
 	}
 };
 
 export const createIndex = async (data) => {
 	const name = data?.name;
-	if (name) {
+	const dbName = data?.dbName;
+	const tbName = data?.tbName;
+	if (name && dbName && tbName) {
 		return await fetch(`${baseURL}index`, {
 			method: 'POST',
 			body: JSON.stringify({
-				name
+				name,
+				dbName,
+				tbName
 			}),
 			headers: {
 				'Content-Type': 'application/json'
